Hide python stack traces if subprocesses fail.
We aren't capturing stderr in the subprocesses, so the standard logging
in conjunction with the child process's stderroutput should be plenty of
information--and indeed the information in the python stack trace
tells us only that there was a child process--which we already knew!

Re-throwing the child process's exit code is almost certainly vanity,
but seems like a good habit.