Remove os.rename(mkstemp()) calls
mkstemp was generating a tempfile on a different volume. rename would
then fail because it doesn't have any cross-volume copy fallback?

check_continue no longer needs a custom exit protocol (which was always
a bit of a hack anyway); the pullme description file is now always
./pullme_description (unless -f was passed) and it's removed if a title
was found.

generate_password_file now manually copies what mkstemp was doing, but
it starts from ~/.githubpass instead of creating a tempfile and
renaming.