Extract subprocess helpers into a module
Mostly this class delegates to the core subprocess module.

It provides check_output, which is not in 2.6's subprocess.

It also provides get_output, which works like check_output but doesn't
throw an exception on a non-0 return status (this is useful to us when
checking git config for settings that may not exist; it returns non-0 in
these cases)